Extract SampleScreenContent
