+++ sample/src/commonMain/kotlin/com/moriatsushi/compose/callable/sample/SampleScreenState.kt
@@ -10,21 +10,19 @@ import kotlinx.coroutines.launch
 internal class SampleScreenState(
     private val coroutineScope: CoroutineScope,
 ) {
-    var selectedTarget by mutableStateOf(CallableTarget.YES_NO_DIALOG)
-        private set
-
     val yesNoDialogState = CallableState<String, Boolean>()
     val confirmDialogState = CallableState<String, Unit>()
 
     var result: String? by mutableStateOf(null)
         private set
 
-    fun call() {
+    fun call(target: CallableTarget) {
         coroutineScope.launch {
-            result = when (selectedTarget) {
-                CallableTarget.YES_NO_DIALOG -> callYesNoDialog()
-                CallableTarget.CONFIRMATION_DIALOG -> callConfirmationDialog()
-            }
+            result =
+                when (target) {
+                    CallableTarget.YES_NO_DIALOG -> callYesNoDialog()
+                    CallableTarget.CONFIRMATION_DIALOG -> callConfirmationDialog()
+                }
         }
     }
 
@@ -36,9 +34,5 @@ internal class SampleScreenState(
     private suspend fun callConfirmationDialog(): String {
         confirmDialogState.call("Please press the Confirm button.")
         return "Confirmed"
-    }
-
-    fun selectTarget(target: CallableTarget) {
-        selectedTarget = target
     }
 }
